Support GitHub Issues and drop b.p.o
Expect gh- issue number in titles.
Remove logic for adding bpo url in the body.
Validate that the GH- issue is indeed an issue, not PR

Closes https://github.com/python/bedevere/issues/418
